Updats from PR review 4
