+++ app/(gcforms)/[locale]/(form filler)/id/[...props]/lib/tests/notifications.vitest.ts
@@ -1,5 +1,5 @@
 import { NotificationsInterval } from "@gcforms/types";
-import { validateNotificationsInterval } from "../notifications";
+import { validateNotificationsInterval } from "@lib/notifications";
 
 const validIntervals = [
   NotificationsInterval.DAY,
