+++ lib/notifications.ts
@@ -1,0 +1,9 @@
+import { NotificationsInterval } from "@gcforms/types";
+
+export const validateNotificationsInterval = (
+  notificationsInterval: number | null | undefined
+): notificationsInterval is NotificationsInterval => {
+  return Object.values(NotificationsInterval).includes(
+    notificationsInterval as NotificationsInterval
+  );
+};
